# docs/apispec.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "MyoHanMeeting Backend",
        "description": "유기묘 입양 서비스 묘한만남의 백엔드 서버 문서입니다.",
        "contact": {
            "email": "bomlee427@gmail.com"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/v1/auth/signin/oauth": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "SNS 회원 로그인",
                "description": "SNS 정보로 로그인합니다.",
                "operationId": "oauthSignIn",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/oauthSignInRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "잘못된 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "400 BAD_REQUEST",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "INVALID PARAMETER",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버측 오류입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "500 INTERNAL_SERVER_ERROR",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "SERVER ERROR",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonResponseDtoTokenDto"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "잘못된 메서드 또는 URI 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "405 METHOD_NOT_ALLOWED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "METHOD_NOT_ALLOWED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/auth/signin/direct": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "이메일 회원 로그인",
                "description": "이메일과 비밀번호로 직접 로그인합니다.",
                "operationId": "directSignIn",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/directSignInRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "잘못된 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "400 BAD_REQUEST",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "INVALID PARAMETER",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버측 오류입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "500 INTERNAL_SERVER_ERROR",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "SERVER ERROR",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonResponseDtoTokenDto"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "잘못된 메서드 또는 URI 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "405 METHOD_NOT_ALLOWED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "METHOD_NOT_ALLOWED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "토큰 리프레시",
                "description": "토큰 리프레시를 요청합니다.",
                "operationId": "refreshToken",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/tokenRefreshRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "잘못된 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "400 BAD_REQUEST",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "INVALID PARAMETER",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버측 오류입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "500 INTERNAL_SERVER_ERROR",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "SERVER ERROR",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonResponseDtoTokenDto"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "허가된 접근이 아닙니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "403 BAD_REQUEST",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "ACCESS DENIED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "인증이 필요한 URI입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "401 UNAUTHORIZED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "AUTHORIZATION REQUIRED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "잘못된 메서드 또는 URI 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "405 METHOD_NOT_ALLOWED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "METHOD_NOT_ALLOWED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "description": "HTTP 상태 코드입니다.",
                        "enum": [
                            "200 OK",
                            "400 BAD_REQUEST",
                            "401 UNAUTHORIZED",
                            "403 FORBIDDEN",
                            "404 NOT_FOUND",
                            "405 METHOD_NOT_ALLOWED",
                            "500 INTERNAL_SERVER_ERROR"
                        ]
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "message": {
                        "type": "string",
                        "description": "응답에 대한 요약을 나타냅니다."
                    },
                    "debugMessage": {
                        "type": "string",
                        "description": "오류에 관한 상세 디버깅 메시지를 나타냅니다."
                    }
                }
            },
...
            "oauthSignInRequest": {
                "type": "object",
                "properties": {
                    "oauthType": {
                        "type": "string"
                    },
                    "oauthId": {
                        "type": "string"
                    }
                }
            },
            "CommonResponseDtoTokenDto": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "200 OK"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "message": {
                        "type": "string",
                        "example": "SUCCESS"
                    },
                    "data": {
                        "$ref": "#/components/schemas/TokenDto"
                    }
                }
            },
            "TokenDto": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string"
                    },
                    "refreshToken": {
                        "type": "string"
                    }
                }
            },
            "directSignInRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "tokenRefreshRequest": {
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}